requirements.txt
streamlit
langchain
langchain-openai
langchain-community
faiss-cpu
PyPDF2
python-dotenv
pydantic
pandas
openpyxl
tiktoken
